pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = ""
authors = ["Eh The Moo Htoo <ehthemoo@gmail.com>", "Ryan Grotzky <ryan.grotzky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
django-cors-headers = "^4.7.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"